evals/requirements.txt
inspect_ai>=0.3.0
python-dotenv>=1.0.0
anthropic>=0.25.0
openai>=1.0.0
